Move first reservable time calculator to services
