+++ tests/test_utils/test_first_reservable_time_helper.py
@@ -11,11 +11,11 @@ from graphene_django_extensions.testing.utils import parametrize_helper
 from tilavarauspalvelu.enums import ReservationStartInterval
 from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
 from tilavarauspalvelu.models import ReservationUnit
-from tilavarauspalvelu.utils.first_reservable_time.first_reservable_time_helper import FirstReservableTimeHelper
-from tilavarauspalvelu.utils.first_reservable_time.first_reservable_time_reservable_time_span_helper import (
+from tilavarauspalvelu.services.first_reservable_time.first_reservable_time_helper import FirstReservableTimeHelper
+from tilavarauspalvelu.services.first_reservable_time.first_reservable_time_reservable_time_span_helper import (
     ReservableTimeSpanFirstReservableTimeHelper,
 )
-from tilavarauspalvelu.utils.first_reservable_time.first_reservable_time_reservation_unit_helper import (
+from tilavarauspalvelu.services.first_reservable_time.first_reservable_time_reservation_unit_helper import (
     ReservationUnitFirstReservableTimeHelper,
 )
 from utils.date_utils import DEFAULT_TIMEZONE
