+++ tilavarauspalvelu/services/first_reservable_time/first_reservable_time_reservable_time_span_helper.py
@@ -8,12 +8,12 @@ from tilavarauspalvelu.enums import ReservationStartInterval
 from tilavarauspalvelu.integrations.opening_hours.time_span_element_utils import (
     override_reservable_with_closed_time_spans,
 )
-from tilavarauspalvelu.utils.first_reservable_time.utils import ReservableTimeOutput
+from tilavarauspalvelu.services.first_reservable_time.utils import ReservableTimeOutput
 
 if TYPE_CHECKING:
     from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
     from tilavarauspalvelu.models import ReservableTimeSpan
-    from tilavarauspalvelu.utils.first_reservable_time.first_reservable_time_reservation_unit_helper import (
+    from tilavarauspalvelu.services.first_reservable_time.first_reservable_time_reservation_unit_helper import (
         ReservationUnitFirstReservableTimeHelper,
     )
