+++ tilavarauspalvelu/services/first_reservable_time/utils.py
